some fixes with expiry time in token'

+++ controllers/scream-controller.js
@@ -8,7 +8,7 @@ const createScream = async (req, res, next) => {
 	const { body } = req.body;
 	const user = req.user;
 	try {
-		const scream = await Scream.create({ body, username: user.username });
+		const scream = await Scream.create({ body, username: user.username, userImage: user.userImage });
 		res.status(201).json({ scream });
 	} catch (err) {
 		const error = new HttpError('Something went wrong !', 500);
@@ -19,19 +19,22 @@ const createScream = async (req, res, next) => {
 const likeScream = async (req, res, next) => {
 	const user = req.user;
 	const screamId = req.params.id;
+	const recipient = await Scream.findOne({ _id: screamId });
 	try {
 		const existingLike = await Like.find({ screamId, username: user.username });
 		// console.log(existingLike);
 		if (existingLike.length === 0) {
 			const like = await Like.create({ screamId, username: user.username });
-			const recipient = await Scream.findOne({ _id: screamId });
-			const newNotification = {
-				type      : 'like',
-				screamId  : screamId,
-				sender    : user.username,
-				recipient : recipient.username
-			};
-			const notification = await Notification.create(newNotification);
+			if (recipient.username !== user.username) {
+				const newNotification = {
+					type        : 'like',
+					screamId    : screamId,
+					sender      : user.username,
+					recipient   : recipient.username,
+					senderImage : user.userImage
+				};
+				const notification = await Notification.create(newNotification);
+			}
 			await Scream.findOne({ _id: screamId }, async function(err, doc) {
 				doc.likeCount = doc.likeCount + 1;
 				await doc.save();
@@ -45,7 +48,9 @@ const likeScream = async (req, res, next) => {
 				doc2.likeCount = doc2.likeCount - 1;
 				await doc2.save();
 			});
-			await Notification.deleteOne({ sender: user.username, screamId: screamId });
+			if (user.username !== recipient.username) {
+				await Notification.deleteOne({ sender: user.username, screamId: screamId });
+			}
 			res.status(200).json({ meassage: 'like removed!' });
 		}
 	} catch (err) {
@@ -68,17 +73,21 @@ const commentOnScream = async (req, res, next) => {
 		const newComment = {
 			body,
 			screamId,
-			username : user.username
+			username  : user.username,
+			userImage : user.userImage
 		};
 		const comment = await Comment.create(newComment);
 		const recipient = await Scream.findById(screamId);
-		const newNotification = {
-			type      : 'comment',
-			screamId  : screamId,
-			sender    : user.username,
-			recipient : recipient.username
-		};
-		const notification = await Notification.create(newNotification);
+		if (recipient.username !== user.username) {
+			const newNotification = {
+				type        : 'comment',
+				screamId    : screamId,
+				sender      : user.username,
+				recipient   : recipient.username,
+				senderImage : user.userImage
+			};
+			const notification = await Notification.create(newNotification);
+		}
 		scream.commentCount = scream.commentCount + 1;
 		scream.save();
 		res.status(201).json({ comment });
